requirements.txt
opencv-python>=4.12.0.88
opencv-python-headless>=4.12.0.88
diffusers>=0.34.0
transformers==4.57.3
tokenizers>=0.20.3
accelerate>=1.8.1
tqdm
imageio
easydict
ftfy
imageio-ffmpeg
scikit-image
loguru
xfuser>=0.4.3
pyloudnorm
decord
xformers==0.0.31
librosa
mediapipe==0.10.9
nvidia-nccl-cu12==2.27.3
# 服务层依赖
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
pyyaml>=6.0
requests>=2.31.0